requirements.txt
requests==2.31.0
ruamel.yaml==0.18.6
typer[all]==0.12.3